Split the code related to options
